src/traffic_info/web_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.1
branca==0.7.2
celery==5.4.0
celery-progress==0.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
contourpy==1.2.1
coverage==6.4.4
coverage-badge==1.1.0
cycler==0.12.1
dbfread==2.0.7
Django==4.2.1
django-cors-headers==4.0.0
django-redis==5.4.0
djangorestframework==3.15.2
docker==7.1.0
fiona==1.9.6
folium==0.16.0
fonttools==4.53.1
freezegun==1.5.1
GDAL==3.6.2
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
geoserver-rest==2.7.0
greenlet==3.0.3
gtfs==0.1.2
gtfs_parser==0.2.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jpholiday==0.1.10
kiwisolver==1.4.5
kombu==5.4.2
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.9.1
networkx==3.3
numpy==1.26.4
osmnx==1.9.3
osmx==0.0.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pdfkit==1.0.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
py==1.11.0
pycapnp==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
redis==5.2.0
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
threadpoolctl==3.5.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
wsproto==1.2.0
xmltodict==0.13.0
xyzservices==2024.4.0
